Switch timer 0 and timer 1
Automatic synchronisation of the OC0B output to XCK will require the
use of the input capture unit, which exists only on timer 1.  Switch
to using OC1B on timer 1 in preparation for this change.

Since timer 0 is only 8 bits, we cannot use OCR0A to produce a 16Hz
interrupt rate; add a software scaler to work around this limitation.

Signed-off-by: Sylvie Barlow <me@sylviebarlow.co.uk>