Decentralise code from the header file
In order to run as many different types of animations as possible it
is important to decentralise any code that is not applicable to all
animations.  For instance code that is designed to light Christmas
displays might not require a concept of frame number.

Allow for this by removing the central concepts of animation and frame
number to the individual animation files, and use an abstraction that
provides only "get next data byte" and "move to next animation"
function".

Signed-off-by: Sylvie Barlow <me@sylviebarlow.co.uk>